Always look up things in the 32bit portion of the registry.
Apparently, the MySQL installation (as of 5.6.21.1 MSI installer) will
create all registry entries in the 32-bit portion of the registry.
When run on a 64-bit R, one has to make sure to access that portion
explicitely, by requesting a 32-bit view of the registry.

Also see http://bugs.mysql.com/bug.php?id=42423.