Consider multiple candidate directories.
Current MySQL installers will install multiple components, and create
registry keys for several of these.  As a consequence, it is incorrect to
assume that any registry node under "MySQL AB" will have a "Location" key,
and even for those who do, some might be connectors or similar instead of
MySQL servers.

So we now postpone the selection of a single directory, and instead do some
more processing using a list of candidate directories.  We try to locate the
dll in one of these, and break if we succeed.