Allow several dbConnect() parameters to be NULL
With the port to Rcpp/C++ (starting with commit b74b697396), the
default value for several parameters of the `dbConnect()` method
changed from `NULL` to `""`. This was probably necessary because a
C++ `std::string` cannot be `NULL`. However, the MySQL C API function
`mysql_real_connect` does accept `NULL` and even requires it for a
password to be read from an option file:

> Then, in the `mysql_real_connect()` call, specify the “no-value”
> value for each parameter to be read from an option file:
> [..]
> - For `passwd`, specify a value of `NULL`. (For the password, a
> value of the empty string in the `mysql_real_connect()` call cannot
> be overridden in an option file, because the empty string indicates
> explicitly that the MySQL account must have an empty password.)
>
> -- [MySQL 5.7 Reference Manual, 22.8.7.54 mysql_real_connect()]
> (https://dev.mysql.com/doc/refman/5.7/en/mysql-real-connect.html)

This patch will again allow `NULL` as an argument for `dbname`,
`username`, `password`, `host`, `unix.socket`, `groups` and
`default.file`. Except for `groups`, which defaults to `"rs-dbi"`,
the default values of these parameters will also be `NULL` again.

This is accomplished by wrapping the `std::string` in the recently
introduced `Rcpp::Nullable` class and checking for `isNull()` in the
call to the MySQL C code.

Because of the use of `Rcpp::Nullable`, the Rcpp dependency now has
a version requirement (>=0.12.1).

Possible alternatives:
- http://stackoverflow.com/questions/341743/c-string-that-can-be-null
- [boost::optional]
(http://www.boost.org/doc/libs/1_59_0/libs/optional/doc/html/index.html)

Fixes rstats-db/RMySQL#55