Use actual connection info for MySQLConnection slots
`MySQLConnection` class slots `host` and `db` were filled with the
`host` and `dbname` arguments passed to `dbConnect()`. However, these
values may be `NULL` when using an option file.

With this patch, the slots will be filled with the values of `host`
and `db` members of the MySQL C API's MYSQL struct created by the
actual connection.